start refactoring combat
